pyproject.toml
[project]
name = "kisanone-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "firebase-admin>=7.1.0",
    "google-genai>=1.36.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]
